[#13] Implement Evaluator class.
